Module rewatch::bsconfig

Structs

PackageSource

PackageSpec

Reason

T

bsconfig.json representation

Warnings

Enums

Error

OneOrMore

Source

Subdirs

Functions

flatten_flags

This flattens string flags

flatten_ppx_flags

Since ppx-flags could be one or more, and could be nested potentiall, this function takes the flags and flattens them outright.

read

Try to convert a bsconfig from a certain path to a bsconfig struct

to_qualified_without_children

to_qualified_without_children takes a tree like structure of dependencies, coming in from bsconfig, and turns it into a flat list. The main thing we extract here are the source folders, and optional subdirs, where potentially, the subdirs recurse or not.