Function rewatch::bsconfig::flatten_ppx_flags

pub fn flatten_ppx_flags(
    root_dir: &String,
    flags: &Option<Vec<OneOrMore<String>>>
) -> Vec<String>

Since ppx-flags could be one or more, and could be nested potentiall, this function takes the flags and flattens them outright.